requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.0
prometheus-fastapi-instrumentator==7.1.0
httpx==0.27.0
websockets==12.0
